refactor: update repository methods for clarity and consistency in naming and return types

+++ src/main/java/com/sprint/mission/discodeit/notification/domain/NotificationRepository.java
@@ -9,5 +9,5 @@ public interface NotificationRepository extends JpaRepository<Notification, UUID
 
     List<Notification> findByReceiverIdAndCheckedFalseOrderByCreatedAtDesc(UUID receiverId);
 
-    long deleteByReceiverId(UUID receiverId);
+    int deleteByReceiverId(UUID receiverId);
 }
